sakai-ng 2 cruds

+++ src/app/demo/components/pages/crud/crud.component.ts
@@ -9,45 +9,22 @@ import { PetService } from '../../../../demo/service/pet.service';
     providers: [MessageService]
 })
 export class CrudComponent implements OnInit {
-
     petDialog: boolean = false;
-
     deletePetDialog: boolean = false;
-
     deletePetsDialog: boolean = false;
 
     pets: Pet[] = [];
-
     pet: Pet = {};
-
     selectedPets: Pet[] = [];
-
     submitted: boolean = false;
-
-    cols: any[] = [];
-
-    statuses: any[] = [];
 
     rowsPerPageOptions = [5, 10, 20];
 
     constructor(private petService: PetService, private messageService: MessageService) { }
 
     ngOnInit() {
+        // Carregar todos os pets
         this.petService.getPets().subscribe(data => this.pets = data);
-
-        this.cols = [
-            { field: 'pet', header: 'Pet' },
-            { field: 'price', header: 'Price' },
-            { field: 'category', header: 'Category' },
-            { field: 'rating', header: 'Reviews' },
-            { field: 'inventoryStatus', header: 'Status' }
-        ];
-
-        this.statuses = [
-            { label: 'INSTOCK', value: 'instock' },
-            { label: 'LOWSTOCK', value: 'lowstock' },
-            { label: 'OUTOFSTOCK', value: 'outofstock' }
-        ];
     }
 
     openNew() {
@@ -72,17 +49,21 @@ export class CrudComponent implements OnInit {
 
     confirmDeleteSelected() {
         this.deletePetsDialog = false;
-        // this.pets = this.pets.filter(val => !this.selectedPets.includes(val));
-        this.petService.deletePet(this.pet.key);
+        this.selectedPets.forEach(selectedPet => {
+            if (selectedPet.key) {
+                this.petService.deletePet(selectedPet.key);
+            }
+        });
         this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pets Deleted', life: 3000 });
         this.selectedPets = [];
     }
 
     confirmDelete() {
         this.deletePetDialog = false;
-        // this.pets = this.pets.filter(val => val.id !== this.pet.id);
-        this.petService.deletePet(this.pet.key);
-        this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pet Deleted', life: 3000 });
+        if (this.pet.key) {
+            this.petService.deletePet(this.pet.key);
+            this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pet Deleted', life: 3000 });
+        }
         this.pet = {};
     }
 
@@ -93,51 +74,31 @@ export class CrudComponent implements OnInit {
 
     savePet() {
         this.submitted = true;
-
-        if (this.pet.name?.trim()) {
-            if (this.pet.id) {
-                // @ts-ignore
-                this.pet.inventoryStatus = this.pet.inventoryStatus.value ? this.pet.inventoryStatus.value : this.pet.inventoryStatus;
-                // this.pets[this.findIndexById(this.pet.id)] = this.pet;
-                this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pet Updated', life: 3000 });
-                this.petService.updatePet(this.pet.key, this.pet);
+        if (this.pet.nome?.trim()) {
+            if (this.pet.key) {
+                // Atualiza pet existente
+                this.petService.updatePet(this.pet.key, this.pet).then(() => {
+                    this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pet Updated', life: 3000 });
+                });
             } else {
-                this.pet.id = this.createId();
+                // Cria novo pet
                 this.petService.createPet(this.pet);
-                // @ts-ignore
-                this.pet.inventoryStatus = this.pet.inventoryStatus ? this.pet.inventoryStatus.value : 'INSTOCK';
-                // this.pets.push(this.pet);
                 this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Pet Created', life: 3000 });
             }
-
-            this.pets = [...this.pets];
             this.petDialog = false;
             this.pet = {};
         }
     }
 
-    findIndexById(id: string): number {
-        let index = -1;
-        for (let i = 0; i < this.pets.length; i++) {
-            if (this.pets[i].id === id) {
-                index = i;
-                break;
-            }
-        }
-
-        return index;
-    }
-
-    createId(): string {
-        let id = '';
-        const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        for (let i = 0; i < 5; i++) {
-            id += chars.charAt(Math.floor(Math.random() * chars.length));
-        }
-        return id;
-    }
-
     onGlobalFilter(table: Table, event: Event) {
         table.filterGlobal((event.target as HTMLInputElement).value, 'contains');
+    }
+
+    // Função para permitir apenas letras nos campos de texto específicos
+    allowLettersOnly(event: KeyboardEvent) {
+        const charCode = event.key.charCodeAt(0);
+        if (!/[a-zA-ZÀ-ÿ\s]/.test(String.fromCharCode(charCode))) {
+            event.preventDefault();
+        }
     }
 }
